Add Helps/Bundle+Extensions.swift Move Bundle extension from FreeAPSApp.swift -> Helpers file to be consistent with other Extensions
