+++ FreeAPS/Sources/Application/FreeAPSApp.swift
@@ -59,23 +59,3 @@ import Swinject
         }
     }
 }
-
-extension Bundle {
-    var releaseVersionNumber: String? {
-        infoDictionary?["CFBundleShortVersionString"] as? String
-    }
-
-    var buildVersionNumber: String? {
-        infoDictionary?["CFBundleVersion"] as? String
-    }
-
-    var buildDate: Date {
-        if let infoPath = Bundle.main.path(forResource: "Info", ofType: "plist"),
-           let infoAttr = try? FileManager.default.attributesOfItem(atPath: infoPath),
-           let infoDate = infoAttr[.modificationDate] as? Date
-        {
-            return infoDate
-        }
-        return Date()
-    }
-}
